refactor(api): Split schemas into 2 files
